--- src/components/App.tsx
@@ -1,11 +1,0 @@
-import { Outlet } from "react-router-dom";
-import Footer from "./Footer";
-
-export default function App() {
-	return (
-		<>
-			<Outlet />
-			<Footer />
-		</>
-	);
-}
