+++ src/components/BasicLayout/LayoutOutlet.tsx
@@ -1,0 +1,11 @@
+import { Outlet } from "react-router-dom";
+import Footer from "./Footer";
+
+export default function LayoutOutlet() {
+  return (
+    <>
+      <Outlet />
+      <Footer />
+    </>
+  );
+}
